requirements.txt
crispy-forms-foundation==0.6.4
dj-config-url==0.1.1
dj-database-url==0.4.2
Django==1.11.6
django-ckeditor==5.3.0
django-crispy-forms==1.7.0
django-js-asset==0.1.1
gunicorn==19.7.1
oauth==1.0.1
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
psycopg2==2.7.3.1
pytz==2017.2
